examples/container-app/microservices/src/product-service/requirements.txt
Flask==3.0.0
opencensus-ext-azure==1.1.13
gunicorn==21.2.0
